Communities Part 1
We can now procedurally generate tables whenever a new game is created. This table (for now) holds the gameid and userid of any user who joins. All joined communities appear in a section on the profile page. Members joined appear for anyone to see. I still have to add: assigning a game manager to a game, and allowing a game manager to edit information about a game (mainly description)