Creation of events.js
I split about 1000 lines of js away from dashboard.html, categorizing it as events.js. This gives me hope we'll be able to do this in a timely fashion.

As of this commit, there is one tiny bug where clicking the delete button on the events tab also opens the details modal, but I will iron that out before pushing.