Game Schedules P1
This is a large development for the project:

Game Schedules can now be created. They are procedurally generated based on the frequency and duration needed. They are only visible to certain categories based on assignment.

Scheduled Events appear uniquely on the calendar, have their own legend section, and show up uniquely in the events tab.

There is a "next scheduled event" modal in a team's page, which changes dynamically to show the next important event for each individual team.

Still need:
-Formatting on next-schedule-event and refactoring size of members/max-size boxes
-Schedule tab shows an overall schedule.
-Mass deleting scheduled events
-New option for "misc" scheduled events. Should have a new color and legend section
-New option for "once" scheduled event, which will only appear one time.